[VigilanceAreas] update table
