--- frontend/src/features/VigilanceArea/components/VigilanceAreasList/Cells/StatusCell.tsx
@@ -1,22 +1,0 @@
-import { Accent, Icon, Tag, THEME } from '@mtes-mct/monitor-ui'
-import styled from 'styled-components'
-
-export function StatusCell({ isDraft }: { isDraft: boolean }) {
-  if (isDraft) {
-    return (
-      <StyledTag accent={Accent.PRIMARY} Icon={Icon.EditUnbordered}>
-        Brouillon
-      </StyledTag>
-    )
-  }
-
-  return (
-    <StyledTag backgroundColor="#D4F0DF" color={THEME.color.mediumSeaGreen}>
-      Publiée
-    </StyledTag>
-  )
-}
-
-const StyledTag = styled(Tag)`
-  font-style: normal;
-`
